feat: add types to task.service.js
